+++ src/components/ChainSelect/ChainSelect.tsx
@@ -1,8 +1,8 @@
 import { MenuItem, makeStyles, Box } from '@material-ui/core'
-import Select from '../../Select/Select'
-import LogoText from '../../LogoText/LogoText'
-import Chain from '../../../models/chain'
-import InputLabel from '../../InputLabel/InputLabel'
+import Select from '../Select/Select'
+import LogoText from '../LogoText/LogoText'
+import Chain from '../../models/chain'
+import InputLabel from '../InputLabel/InputLabel'
 import SelectedIcon from '../../../assets/images/selected_icon.svg'
 
 interface Props {
