+++ src/pages/Liquidity/AddLiquidity.tsx
@@ -8,7 +8,7 @@ import { TYPE } from 'theme'
 import { ReactComponent as ArrowLeft } from '../../assets/images/arrow_left.svg'
 import { ChainList, CurrencyList } from 'data/dummyData'
 import Currency from 'models/currency'
-import ChainSelect from 'components/swap/ChainSelectPanel/ChainSelect'
+import ChainSelect from 'components/ChainSelect/ChainSelect'
 import Chain from 'models/chain'
 import Button from 'components/Button/Button'
 import OutlineButton from 'components/Button/OutlineButton'
