+++ src/pages/Liquidity/ArbitrageModal.tsx
@@ -4,7 +4,7 @@ import { Box, styled, useTheme } from '@material-ui/core'
 import Button from 'components/Button/Button'
 import OutlineButton from 'components/Button/OutlineButton'
 import Input from 'components/Input/Input'
-import ChainSelect from 'components/swap/ChainSelectPanel/ChainSelect'
+import ChainSelect from 'components/ChainSelect/ChainSelect'
 import { ChainList } from 'data/dummyData'
 import Chain from 'models/chain'
 import AppBody from 'pages/AppBody'
